dev_requirements.txt
distribute==0.6.24
django-hudson==0.9.1
pylint==0.25.1